requirements.txt
absl-py==0.11.0
aiohttp==3.7.3
albumentations==0.5.2
alembic==1.4.1
async-timeout==3.0.1
atomicwrites==1.4.0
attrs==20.3.0
azure-core==1.10.0
azure-storage-blob==12.7.1
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
click==7.1.2
cloudpickle==1.6.0
colorama==0.4.4
cryptography==3.3.1
cycler==0.10.0
databricks-cli==0.14.1
decorator==4.4.2
docker==4.4.1
entrypoints==0.3
Flask==1.1.2
fsspec==0.8.5
future==0.18.2
gitdb==4.0.5
GitPython==3.1.12
google-auth==1.24.0
google-auth-oauthlib==0.4.2
grpcio==1.35.0
idna==2.10
imageio==2.9.0
imgaug==0.4.0
iniconfig==1.1.1
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.0
kiwisolver==1.3.1
Mako==1.1.4
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.4
mlflow==1.13.1
msrest==0.6.21
multidict==5.1.0
networkx==2.5
numpy==1.20.0
oauthlib==3.1.0
opencv-python==4.5.1.48
opencv-python-headless==4.5.1.48
packaging==20.9
pandas==1.2.1
Pillow==8.1.0
pluggy==0.13.1
prometheus-client==0.9.0
prometheus-flask-exporter==0.18.1
protobuf==3.14.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==2.4.7
pytest==6.2.2
python-dateutil==2.8.1
python-editor==1.0.4
pytorch-lightning==1.1.8
pytz==2021.1
PyWavelets==1.1.1
pywin32==227
PyYAML==5.3.1
querystring-parser==1.2.4
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7
scikit-image==0.18.1
scikit-learn==0.24.1
scipy==1.6.0
Shapely==1.7.1
six==1.15.0
smmap==3.0.5
SQLAlchemy==1.3.23
sqlparse==0.4.1
tabulate==0.8.7
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
threadpoolctl==2.1.0
tifffile==2021.2.1
toml==0.10.2
torch==1.7.1+cu101
torchaudio==0.7.2
torchvision==0.8.2+cu101
tqdm==4.56.0
typing-extensions==3.7.4.3
urllib3==1.26.3
waitress==1.4.4
websocket-client==0.57.0
Werkzeug==1.0.1
wincertstore==0.2
yarl==1.6.3

python-dotenv~=0.15.0